// Facilitei-Front/src/types/api.ts
export interface Endereco {
  rua: string;
  numero: string;
  bairro: string;
  cidade: string;
  estado: string;
  cep: string;
}

export interface Trabalhador {
  id: string;
  nome: string;
  email: string;
  senha?: string;
  avatarUrl: string;
  telefone: string;
  endereco: Endereco;
  disponibilidade: string;
  notaTrabalhador: number;
  servicos: TipoServico[];
  servicoPrincipal: TipoServico;
  sobre?: string;
}

export interface Cliente {
  id: string;
  nome: string;
  email: string;
  senha?: string;
  telefone: string;
  avatarUrl: string;
  notaCliente: number;
  endereco: Endereco;
}

export type TipoServico =
  // Construção e Reformas
  | "PEDREIRO"
  | "ELETRICISTA"
  | "ENCANADOR"
  | "PINTOR"
  | "GESSEIRO"
  | "AZULEJISTA"
  | "MARCENEIRO"
  | "SERRALHEIRO"
  // Serviços Domésticos
  | "DIARISTA"
  | "FAXINEIRA"
  | "PASSADEIRA"
  | "COZINHEIRA"
  | "CUIDADOR_IDOSOS"
  | "BABA"
  // Serviços Técnicos
  | "TECNICO_INFORMATICA"
  | "TECNICO_ELETRODOMESTICOS"
  | "INSTALADOR_AR_CONDICIONADO"
  | "TECNICO_REDES"
  | "INSTALADOR_CAMERAS"
  // Jardinagem e Exteriores
  | "JARDINEIRO"
  | "PAISAGISTA"
  | "DEDETIZADOR"
  | "LIMPADOR_PISCINA"
  // Educação e Aulas Particulares
  | "REFORCO_ESCOLAR"
  | "AULAS_INGLES"
  | "AULAS_MUSICA"
  | "AULAS_INFORMATICA"
  | "PERSONAL_TRAINER";
// ...
export type CategoriaGrupo =
  | "Construção e Reformas"
  | "Serviços Domésticos"
  | "Serviços Técnicos"
  | "Jardinagem e Exteriores"
  | "Educação e Aulas";

export const serviceCategories: Record<CategoriaGrupo, TipoServico[]> = {
  "Construção e Reformas": [
    "PEDREIRO",
    "ELETRICISTA",
    "ENCANADOR",
    "PINTOR",
    "GESSEIRO",
    "AZULEJISTA",
    "MARCENEIRO",
    "SERRALHEIRO",
  ],
  "Serviços Domésticos": [
    "DIARISTA",
    "FAXINEIRA",
    "PASSADEIRA",
    "COZINHEIRA",
    "CUIDADOR_IDOSOS",
    "BABA",
  ],
  "Serviços Técnicos": [
    "TECNICO_INFORMATICA",
    "TECNICO_ELETRODOMESTICOS",
    "INSTALADOR_AR_CONDICIONADO",
    "TECNICO_REDES",
    "INSTALADOR_CAMERAS",
  ],
  "Jardinagem e Exteriores": [
    "JARDINEIRO",
    "PAISAGISTA",
    "DEDETIZADOR",
    "LIMPADOR_PISCINA",
  ],
  "Educação e Aulas": [
    "REFORCO_ESCOLAR",
    "AULAS_INGLES",
    "AULAS_MUSICA",
    "AULAS_INFORMATICA",
    "PERSONAL_TRAINER",
  ],
};

// Helper para pegar todos os nomes dos grupos
export const allCategoryGroups = Object.keys(
  serviceCategories
) as CategoriaGrupo[];